revert import-export refactoring
