+++ src/molecules/ReadMore.jsx
@@ -1,8 +1,7 @@
 import PropTypes from 'prop-types'
 import React from 'react'
 import { css } from 'glamor'
-import View from '../atoms/View'
-import Text from '../atoms/Text'
+import { View, Text } from '../'
 
 /**
  *   A ReadMore is a simple container, to show / (hide) content. It will automatically decide whether to show the `read more` link or not.
