+++ stories/DropdownStory.jsx
@@ -1,11 +1,13 @@
 import React from 'react'
+import {
+  View,
+  ThemeProvider,
+  Dropdown,
+  ResourceProvider,
+  Button,
+} from '../src/'
 import { css } from 'glamor'
 import Text from '../src/atoms/Text'
-import ThemeProvider from '../src/behaviour/ThemeProvider'
-import ResourceProvider from '../src/behaviour/ResourceProvider'
-import View from '../src/atoms/View'
-import Button from '../src/molecules/Button'
-import Dropdown from '../src/organisms/Dropdown'
 
 const styles = {
   title: css({
