+++ stories/FormStory.jsx
@@ -1,22 +1,24 @@
 import React from 'react'
 import { action } from '@storybook/addon-actions'
-import ThemeProvider from '../src/behaviour/ThemeProvider'
-import View from '../src/atoms/View'
-import Text from '../src/atoms/Text'
-import List from '../src/molecules/List/List'
-import ResourceProvider from '../src/behaviour/ResourceProvider'
-import ListItem from '../src/molecules/List/ListItem'
-import TitleBar from '../src/organisms/TitleBar'
-import SquareIconButton from '../src/molecules/SquareIconButton'
-import SimpleLayout from '../src/layouts/SimpleLayout'
-import Form from '../src/molecules/Form/Form'
-import Card from '../src/molecules/Card/Card'
-import TextInput from '../src/molecules/TextInput'
-import PhoneInput from '../src/molecules/PhoneInput'
-import RadioButtonSet from '../src/molecules/RadioButtonSet'
-import RadioButton from '../src/molecules/RadioButton'
-import Button from '../src/molecules/Button'
-import Checkbox from '../src/molecules/Checkbox'
+import {
+  Button,
+  Card,
+  Checkbox,
+  Form,
+  List,
+  ListItem,
+  PhoneInput,
+  ResourceProvider,
+  RadioButtonSet,
+  RadioButton,
+  SimpleLayout,
+  SquareIconButton,
+  Text,
+  TextInput,
+  TitleBar,
+  ThemeProvider,
+  View,
+} from '../src'
 
 export default class FormStory extends React.Component {
   render() {
